#19 refactor: 도메인형 패키지 구조로 변경
