+++ src/main/java/com/hanpyeon/academyapi/account/dto/RegisterMemberTotalDto.java
@@ -1,4 +1,4 @@
-package com.hanpyeon.academyapi.dto;
+package com.hanpyeon.academyapi.account.dto;
 
 import com.hanpyeon.academyapi.security.Role;
 import jakarta.validation.constraints.NotBlank;
@@ -10,11 +10,12 @@ import org.hibernate.validator.constraints.Range;
 import java.time.LocalDateTime;
 
 @Builder
-public record RegisterMemberDto(
+public record RegisterMemberTotalDto (
         @NotBlank String name,
         @Range(min = 0, max = 11) Integer grade,
         @NotBlank @Pattern(regexp = "^[0-9]+$") String phoneNumber,
         @NotNull Role role,
+        @NotNull LocalDateTime registerDate,
         String password
 ){
 }
