+++ src/main/java/com/hanpyeon/academyapi/account/repository/MemberRepository.java
@@ -1,10 +1,8 @@
-package com.hanpyeon.academyapi.repository;
+package com.hanpyeon.academyapi.account.repository;
 
-import com.hanpyeon.academyapi.entity.Member;
+import com.hanpyeon.academyapi.account.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 
 @Repository
