+++ src/main/java/com/hanpyeon/academyapi/account/service/RegisterService.java
@@ -1,8 +1,8 @@
-package com.hanpyeon.academyapi.service;
+package com.hanpyeon.academyapi.account.service;
 
-import com.hanpyeon.academyapi.dto.RegisterMemberDto;
-import com.hanpyeon.academyapi.dto.RegisterMemberTotalDto;
-import com.hanpyeon.academyapi.mapper.RegisterMapper;
+import com.hanpyeon.academyapi.account.dto.RegisterMemberDto;
+import com.hanpyeon.academyapi.account.dto.RegisterMemberTotalDto;
+import com.hanpyeon.academyapi.account.mapper.RegisterMapper;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
